+++ VersionHistory.md
@@ -7,6 +7,7 @@ Describe changes here when they're committed to the `master` branch. Move them t
 Prefix the description of the change with `[major]`, `[minor]`, or `[patch]` in accordance with [Semantic Versioning](https://semver.org/).
 
 * [patch] Trigger detection should support `publish-` tags.
+* [patch] Update `xmldocmd`.
 
 ## Released